lab4 -complete all tests
